+++ app/src/models/endpoint.model.js
@@ -17,19 +17,7 @@ const Endpoint = new Schema({
         path: { type: String, required: true, trim: true },
         url: { type: String, required: true, trim: true },
         method: { type: String, required: true, trim: true },
-        microservice: { type: String, required: false, trim: true },
-        filters: [{
-            name: { type: String, required: true, trim: true },
-            path: { type: String, required: true, trim: true },
-            condition: {
-                type: String, required: true, trim: true, default: 'AND'
-            },
-            method: { type: String, required: true, trim: true },
-            pathRegex: { type: RegExp, required: true },
-            pathKeys: [{ type: String, trim: true }],
-            params: Schema.Types.Mixed,
-            compare: Schema.Types.Mixed,
-        }],
+        microservice: { type: String, required: false, trim: true }
     }],
     version: { type: Number, required: true },
     updatedAt: { type: Date, default: Date.now, required: true },