+++ app/src/services/dispatcher.service.js
@@ -6,9 +6,7 @@ const url = require('url');
 const EndpointNotFound = require('errors/endpointNotFound');
 const NotAuthenticated = require('errors/notAuthenticated');
 const NotApplicationKey = require('errors/notApplicationKey');
-const FilterError = require('errors/filterError');
 const pathToRegexp = require('path-to-regexp');
-const requestPromise = require('request-promise');
 const fs = require('fs');
 
 const ALLOWED_HEADERS = [
@@ -126,33 +124,6 @@ class Dispatcher {
         return false;
     }
 
-    static checkValidRedirects(redirects, filters) {
-        logger.debug('Checking redirects with filters');
-        const validRedirects = [];
-        for (let i = 0, { length } = redirects; i < length; i++) {
-            const redirect = redirects[i];
-            let valid = true;
-            if (redirect.filters) {
-                for (let j = 0, lengthRF = redirect.filters.length; j < lengthRF; j++) {
-                    const filterValue = Dispatcher.searchFilterValue(redirect.filters[j], filters);
-                    if (!filterValue || !Dispatcher.checkCompare(redirect.filters[j].compare, filterValue, redirect.filters[j].condition)) {
-                        logger.debug('Not valid filter');
-                        valid = false;
-                        break;
-                    }
-                    if (!redirect.data) {
-                        redirect.data = {};
-                    }
-                    redirect.data[redirect.filters[j].name] = filterValue;
-                }
-            }
-            if (valid) {
-                validRedirects.push(redirect);
-            }
-        }
-        return validRedirects;
-    }
-
     static cloneEndpoint(endpoint) {
         const newObject = { ...endpoint };
         newObject.redirect = [];
@@ -160,67 +131,6 @@ class Dispatcher {
             newObject.redirect.push({ ...endpoint.redirect[i] });
         }
         return newObject;
-    }
-
-    static async checkFilters(sourcePath, endpoint) {
-        logger.debug('[DispatcherService - checkFilters] Checking filters in endpoint', endpoint);
-        const newEndpoint = Dispatcher.cloneEndpoint(endpoint);
-        let filters = [];
-        for (let i = 0, { length } = newEndpoint.redirect; i < length; i++) {
-            if (newEndpoint.redirect[i].filters) {
-                filters = filters.concat(newEndpoint.redirect[i].filters);
-            }
-        }
-        if (!filters || filters.length === 0) {
-            logger.debug('[DispatcherService - checkFilters] Doesn\'t contain filters. All redirect are valid');
-            return newEndpoint;
-        }
-        logger.debug('[DispatcherService - checkFilters] Obtaining data to check filters');
-        const promisesRequest = [];
-        let filter = null;
-        let path = null;
-        for (let i = 0, { length } = filters; i < length; i++) {
-            filter = filters[i];
-            path = await Dispatcher.buildPathFilter(sourcePath, filter, newEndpoint);
-            const request = await Dispatcher.getRequest({
-                request: {
-                    url: path,
-                    method: filter.method,
-                    body: {},
-                },
-            });
-            logger.debug('[DispatcherService - checkFilters] Config request', request);
-            request.configRequest.json = true;
-            promisesRequest.push(requestPromise(request.configRequest));
-
-        }
-        if (!promisesRequest || promisesRequest.length === 0) {
-            return newEndpoint;
-        }
-        try {
-            logger.debug('[DispatcherService - checkFilters] Doing requests');
-            const results = await Promise.all(promisesRequest);
-            // TODO: Add support for several filter in each newEndpoint
-            for (let i = 0, { length } = results; i < length; i++) {
-                if (results[i].statusCode === 200) {
-                    filters[i].result = {
-                        data: results[i].body,
-                        correct: true,
-                    };
-                } else {
-                    filters[i].result = {
-                        correct: false,
-                    };
-                }
-            }
-            logger.debug('[DispatcherService - checkFilters] Checking valid filters');
-            const validRedirects = Dispatcher.checkValidRedirects(newEndpoint.redirect, filters);
-            newEndpoint.redirect = validRedirects;
-        } catch (err) {
-            logger.error(err);
-            throw new FilterError('Error resolving filters');
-        }
-        return newEndpoint;
     }
 
     static getHeadersFromRequest(headers) {
@@ -280,7 +190,7 @@ class Dispatcher {
         logger.info(`[DispatcherService - getRequest] Searching endpoint where redirect url ${ctx.request.url}
             and method ${ctx.request.method}`);
         const parsedUrl = url.parse(ctx.request.url);
-        let endpoint = await Dispatcher.getEndpoint(parsedUrl.pathname, ctx.request.method);
+        const endpoint = await Dispatcher.getEndpoint(parsedUrl.pathname, ctx.request.method);
 
         if (!endpoint) {
             throw new EndpointNotFound(`${parsedUrl.pathname} not found`);
@@ -298,7 +208,6 @@ class Dispatcher {
             throw new NotApplicationKey('Required app_key');
         }
         let redirectEndpoint = null;
-        endpoint = await Dispatcher.checkFilters(parsedUrl.pathname, endpoint);
         if (endpoint && endpoint.redirect.length === 0) {
             logger.error('[DispatcherService - getRequest] No redirects exist');
             throw new EndpointNotFound(`${parsedUrl.pathname} not found`);